Added default repo provider setting
