+++ src/configure/configurers/remoteGitHubWorkflowConfigurer.ts
@@ -6,6 +6,7 @@ import { ArmRestClient } from '../clients/azure/armRestClient';
 import { AzureResourceClient } from "../clients/azure/azureResourceClient";
 import { GithubClient } from '../clients/github/githubClient';
 import { TemplateServiceClient } from '../clients/github/TemplateServiceClient';
+import { createOrUpdateGithubSecret } from "../helper/commonHelper";
 import { LocalGitRepoHelper } from '../helper/LocalGitRepoHelper';
 import { MustacheHelper } from '../helper/mustacheHelper';
 import { telemetryHelper } from '../helper/telemetryHelper';
@@ -136,7 +137,7 @@ export class RemoteGitHubWorkflowConfigurer extends LocalGitHubWorkflowConfigure
                             title: Messages.settingUpGithubSecrets
                         },
                         async () => {
-                            await this.githubClient.createOrUpdateGithubSecret(secretKey, secretValue);
+                            await createOrUpdateGithubSecret(secretKey, secretValue);
                         }
                     );
                     this.secrets[asset.id] = secretKey;
