+++ src/configure/resources/tracePoints.ts
@@ -24,5 +24,5 @@ export class TracePoints {
     public static InitializeDynamicValidation = 'InitializeDynamicValidation';
     public static SetInputControlValueFromRepoAnalysisResult = 'setInputControlValueFromRepoAnalysisResult';
     public static GetTemplateFile = 'GetTemplateFile';
-    public static GetGitHubDetailsFailed = 'GetGitHubDetailsFailed';
+    public static NoGitHubOrganizationExists = 'NoGitHubOrganizationExists';
 }
